net: Allow vlan offload to work with XDP programs
VLAN acceleration enabled in hardware means hardware handles stripping
the vlan header with the tag set in the skb. Any XDP progrms running
for the netdevice, see the modified packet without the vlan header.
However, the VLAN tag is often essential information that a program
needs to make, for example, a forwarding decision.

One solution is to require vlan acceleration to be disabled if an
XDP program relies on the vlan data, but that means XDP program has to
process the vlan header to get the vlan tag, which is counter to the
very intent of XDP - to not have to perform any checks that can be
deferred to other places.

This patch adds the vlan tag to the xdp context, xdp_md, and the kernel
side version of it, xdp_buff. A follow on patch can look at support for
vlan acceleration in the Tx path - allowing H/W to add the vlan header
instead of doing this in the XDP program.

I am aware of the earlier effort regarding xdp meta data. IMHO, vlan
tags are NOT meta data like checksums or hashes, but rather core data
essential to a networking setup. As such it warrants a more direct
place in the XDP structures.

Signed-off-by: David Ahern <dahern@digitalocean.com>