xdp: Propagate ip_summed through XDP frames and buffs
skb's have ip_summed for drivers to indicate to the stack that
the H/W has validated the L3 and L4 checksums of the packet. Add
the same to xdp_buff and xdp_frame for VMs (via tap devices) and
containers (via veth) to get the same benefit and avoid the do_csum
or csum_partial overhead when using XDP redirect.

frame_sz only needs 17-bits; steal 2 for ip_summed.

Signed-off-by: David Ahern <dsahern@gmail.com>